manga-python/requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
arabic-reshaper==2.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diffimg==0.2.3
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.6.0
flatbuffers==24.3.7
fqdn==1.5.1
fsspec==2024.2.0
fugashi==1.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.21.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==7.0.2
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jaconv==0.3.4
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mecab-python3==1.0.8
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
onnx==1.15.0
onnx-simplifier==0.4.36
onnxruntime==1.17.1
opencv-python==4.9.0.80
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyperclip==1.8.2
pyreadline3==3.4.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
Send2Trash==1.8.2
sentry-sdk==1.41.0
setproctitle==1.3.3
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
torchaudio==2.2.1
torchsummary==1.5.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.38.2
trdg==1.8.0
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
tzdata==2024.1
unidic-lite==1.0.8
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.28.0
wandb==0.16.4
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wikipedia==1.4.0
zipp==3.18.0
